authentication added

+++ backend/controllers/todoController.ts
@@ -10,16 +10,18 @@ import {
 
 export const getTodo = async (req: Request, res: Response) => {
     try {
+        const userId = req.user?.id;
+
         const { pageIdx = 1, limit = 5 } = paginationValidationSchema.parse(
             req.query
         );
         const todoList = await todoModel
-            .find({})
+            .find({ userId })
             .sort({ createdAt: -1 })
             .skip((pageIdx - 1) * limit)
             .limit(limit);
 
-        const totalDocs = await todoModel.find().count();
+        const totalDocs = await todoModel.find({ userId }).count();
         const hasprev = pageIdx > 1;
         const hasNext = pageIdx * limit < totalDocs;
 
@@ -39,6 +41,7 @@ export const saveTodo = async (req: Request, res: Response) => {
         const data = await todoModel.create({
             text: validateSave.parse(req.body).text,
             date: new Date(),
+            userId: req.user?.id,
         });
         return res.status(201).json({ status: true, todo: data });
     } catch (err) {
@@ -57,14 +60,19 @@ export const saveTodo = async (req: Request, res: Response) => {
 export const updateTodo = async (req: Request, res: Response) => {
     try {
         const { _id, text } = validateUpdate.parse(req.body);
-
-        const result = await todoModel.findByIdAndUpdate(
-            _id,
-            { text: text },
-            { new: true }
-        );
-        if (result) return res.status(200).json({ status: true });
-        else return res.status(200).json({ status: false });
+        const userId = req.user?.id;
+        const todo = await todoModel.findOne({ _id, userId });
+        if (todo) {
+            const result = await todoModel.findByIdAndUpdate(
+                _id,
+                { text: text },
+                { new: true }
+            );
+            return res.status(200).json({ status: true });
+        } else
+            return res
+                .status(401)
+                .json({ status: false, error: "unauthorised" });
     } catch (err) {
         console.log(err);
         if (err instanceof z.ZodError) {
@@ -81,8 +89,15 @@ export const updateTodo = async (req: Request, res: Response) => {
 export const deleteTodo = async (req: Request, res: Response) => {
     try {
         const { _id }: { _id: string } = validateDelete.parse(req.body);
-        await todoModel.findByIdAndDelete(_id);
-        return res.status(200).json({ status: true });
+        const userId = req.user?.id;
+        const todo = await todoModel.findOne({ _id, userId });
+        if (todo) {
+            await todoModel.findByIdAndDelete(_id);
+            return res.status(200).json({ status: true });
+        } else
+            return res
+                .status(401)
+                .json({ status: false, error: "unauthorised" });
     } catch (err) {
         console.log(err);
         if (err instanceof z.ZodError) {
